[Feature #154239873] Cache nutritional values dictionaries
Previously when the user viewed nutritional values plan overview the
load time was long because nutritional values had to be calculated
for each and every plan.

This commit will fix the overall performance of the application
since the results will be saved in the cache and retrieved if they
exist.